+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -19,6 +19,7 @@ CHANGELOG
  * Removed `SecurityUserValueResolver`, use `UserValueResolver` instead
  * Removed `routing.loader.service`.
  * Service route loaders must be tagged with `routing.route_loader`. 
+ * Added `slugger` service and `SluggerInterface` alias
 
 4.4.0
 -----